\name{flowDensity-methods}
\docType{methods}
\alias{flowDensity-methods}
\alias{flowDensity,CellPopulation,ANY,logical,missing-method}
\alias{flowDensity,GatingHierarchy,ANY,logical,missing-method}
\alias{flowDensity,flowFrame,ANY,logical,missing-method}
\title{ Methods for Function \code{flowDensity} in Package \pkg{flowDensity} }
\description{
  Methods for function \code{flowDensity} in package \pkg{flowDensity} 
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "CellPopulation", channels = "ANY", position = "logical")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "GatingHierarchy", channels = "ANY", position = "logical",node="ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "flowFrame", channels = "ANY", position = "logical")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
